array: fix up types in vlc_array_*()
 - use size_t, ssize_t,
 - handle const types where applicable.

+++ include/vlc_fingerprinter.h
@@ -61,7 +61,7 @@ static inline void fingerprint_request_Delete( fingerprint_request_t *p_f )
 {
     vlc_gc_decref( p_f->p_item );
     free( p_f->results.psz_fingerprint );
-    for( int i = 0; i < vlc_array_count( & p_f->results.metas_array ); i++ )
+    for( size_t i = 0; i < vlc_array_count( & p_f->results.metas_array ); i++ )
         vlc_meta_Delete( (vlc_meta_t *) vlc_array_item_at_index( & p_f->results.metas_array, i ) );
     free( p_f );
 }
